Move a bunch of grid collection view stuff into a controller
