+++ MIDIchlorians/MIDIchlorians/GridCollectionView.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class GridCollectionView: UICollectionView, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+class GridCollectionView: UICollectionView {
     var mode: Mode = .Playing {
         didSet {
             if mode == .Playing, let selectedPad = selectedPad {
@@ -16,62 +16,9 @@ class GridCollectionView: UICollectionView, UICollectionViewDataSource, UICollec
             }
         }
     }
-    internal var currentSession: Session?
-    internal var currentPage = 0
 
-    private let reuseIdentifier = "cell"
     private let audioManager = AudioManager()
     internal var selectedPad: IndexPath?
-
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return Config.numberOfColumns
-    }
-
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return Config.numberOfRows
-    }
-
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath)
-        -> UICollectionViewCell {
-            guard let cell = collectionView.dequeueReusableCell(
-                withReuseIdentifier: reuseIdentifier,
-                for: indexPath as IndexPath) as? GridCollectionViewCell else {
-                    return GridCollectionViewCell()
-            }
-            cell.rowNumber = indexPath.section
-            cell.columnNumber = indexPath.item
-            cell.setAppearance()
-            return cell
-    }
-
-    func collectionView(_ collectionView: UICollectionView,
-                        layout collectionViewLayout: UICollectionViewLayout,
-                        sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let totalLength = self.frame.width
-        let itemsPerRow = CGFloat(collectionView.numberOfItems(inSection: indexPath.section))
-        let insetLength = Config.ItemInsets.left * (itemsPerRow + 1)
-        let availableLength = totalLength - insetLength
-        let itemLength = availableLength / itemsPerRow
-        return CGSize(width: itemLength, height: itemLength)
-    }
-
-    func collectionView(_ collectionView: UICollectionView,
-                        layout collectionViewLayout: UICollectionViewLayout,
-                        minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return Config.ItemInsets.left
-    }
-
-    func collectionView(_ collectionView: UICollectionView,
-                        layout collectionViewLayout: UICollectionViewLayout,
-                        minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return Config.ItemInsets.top
-    }
-
-    func collectionView(_ collectionView: UICollectionView,
-                        layout collectionViewLayout: UICollectionViewLayout,
-                        insetForSectionAt section: Int) -> UIEdgeInsets {
-        return Config.SectionInsets
-    }
 
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         for touch in touches {
@@ -160,30 +107,5 @@ class GridCollectionView: UICollectionView, UICollectionViewDataSource, UICollec
 //                return
 //            }
 //        }
-    }
-}
-
-extension GridCollectionView: SampleTableDelegate {
-    func sampleTable(_: UITableView, didSelect sample: String) {
-        guard let row = self.selectedPad?.section, let col = self.selectedPad?.row else {
-            return
-        }
-        self.currentSession?.addAudio(page: self.currentPage, row: row, col: col, audioFile: sample)
-    }
-}
-
-extension GridCollectionView: AnimationTableDelegate {
-    func animationTable(_: UITableView, didSelect animation: String) {
-        guard let indexPath = self.selectedPad else {
-            return
-        }
-
-        guard let animationSequence = AnimationTypes.getAnimationSequenceForAnimationType(
-            animationTypeName: animation, indexPath: indexPath) else {
-                return
-        }
-
-        self.currentSession?.addAnimation(
-            page: self.currentPage, row: indexPath.section, col: indexPath.row, animation: animationSequence)
     }
 }
